test: refactor some for loop tests into pytest parametrize (#1192)
Refactor some test cases with "for" loop or `subTest` into pytest's `parametrize` fixture functions.
